requirements.txt
Django==2.0
mysqlclient==1.3.12
celery==4.1.0
python-decouple==3.1
dj-database-url==0.4.2
django-debug-toolbar==1.9.1
requests==2.18.4
PyGithub==1.35
